feat: implementa DTO para Produto com validação, enum Categoria e melhorias no controller: conversão no Entity --- DTO , integração de ENUM Categoria ao modelo, ajusta endpoints para retorno com DTO , tratamento básico para produto não encontrado

+++ src/main/java/com/pedro/ambarpastelaria/controller/ProdutoController.java
@@ -1,5 +1,7 @@
 package com.pedro.ambarpastelaria.controller;
 
+import com.pedro.ambarpastelaria.DTO.ProdutoDTO;
+import com.pedro.ambarpastelaria.model.Categoria;
 import com.pedro.ambarpastelaria.model.Produto;
 import com.pedro.ambarpastelaria.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,41 +18,72 @@ public class ProdutoController
     @Autowired
     private ProdutoRepository repository;
 
+    private ProdutoDTO converterParaDTO(Produto produto) {
+        ProdutoDTO dto = new ProdutoDTO();
+
+        dto.setNome(produto.getNome());
+        dto.setPreco(produto.getPreco());
+        dto.setDescricao(produto.getDescricao());
+        dto.setCategoria(produto.getCategoria().name());
+
+        return dto;
+    }
+
     @PostMapping
-    public Produto salvar(@RequestBody @Valid Produto produto)
+    public ProdutoDTO salvar(@RequestBody @Valid ProdutoDTO dto)
     {
-        return repository.save(produto);
+        Produto produto = new Produto();
+
+        produto.setNome(dto.getNome());
+        produto.setPreco(dto.getPreco());
+        produto.setDescricao(dto.getDescricao());
+        produto.setCategoria(Categoria.valueOf(dto.getCategoria().toUpperCase()));
+
+        Produto salvo = repository.save(produto);
+        return converterParaDTO(salvo);
     }
     @GetMapping
-    public List<Produto> listar()
+    public List<ProdutoDTO> listar()
     {
-        return repository.findAll();
+        return repository.findAll()
+                .stream()
+                .map(this::converterParaDTO)
+                .toList();
     }
     @GetMapping("/{id}")
-    public Produto buscarPorId(@PathVariable Long id)
+    public ProdutoDTO buscarPorId(@PathVariable Long id)
     {
-        return repository.findById(id).orElse(null);
+        Produto produto = repository.findById(id).orElse(null);
+
+        if (produto == null) {
+            throw new RuntimeException("Produto não encontrado");
+        }
+        return converterParaDTO(produto);
     }
 
     /*Metodo para atualizar produto. */
     @PutMapping("/{id}")
-    public Produto atualizar(@PathVariable Long id, @RequestBody Produto produtoAtualizado)
+    public ProdutoDTO atualizar(@PathVariable Long id, @RequestBody @Valid ProdutoDTO dto)
     {
         Produto produto = repository.findById(id).orElse(null);
 
-        if(produto == null){
-            return null;
+        if (produto == null) {
+            throw new RuntimeException("Produto não encontrado");
         }
-        produto.setNome(produtoAtualizado.getNome());
-        produto.setPreco(produtoAtualizado.getPreco());
-        produto.setDescricao(produtoAtualizado.getDescricao());
-        produto.setCategoria(produtoAtualizado.getCategoria());
 
-        return repository.save(produto);
+        produto.setNome(dto.getNome());
+        produto.setPreco(dto.getPreco());
+        produto.setDescricao(dto.getDescricao());
+        produto.setCategoria(Categoria.valueOf(dto.getCategoria().toUpperCase()));
+        Produto atualizado = repository.save(produto);
+        return converterParaDTO(atualizado);
     }
 
     @DeleteMapping("/{id}")
     public void deletar(@PathVariable Long id){
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Produto não encontrado");
+        }
         repository.deleteById(id);
     }
 }
